pyproject.toml
[project]
name = "mail-organizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=0.1.0",
    "groq>=1.0.0",
    "huggingface-hub>=1.3.1",
    "msal>=1.34.0",
    "pillow>=12.1.0",
    "pyngrok>=7.5.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
